refactor: remove superfluous widget folders
