--- packages/dm-core-plugins/src/form/widgets/CheckboxWidget/index.ts
@@ -1,2 +1,0 @@
-export { default } from './CheckboxWidget'
-export * from './CheckboxWidget'
