--- packages/dm-core-plugins/src/form/widgets/TextWidget/index.ts
@@ -1,2 +1,0 @@
-export { default } from './TextWidget'
-export * from './TextWidget'
