+++ packages/dm-core-plugins/src/form/widgets/TextareaWidget.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 
-import { TextField, Icon } from '@equinor/eds-core-react'
+import { Icon, TextField } from '@equinor/eds-core-react'
 
 import { error_filled } from '@equinor/eds-icons'
-import { TWidget } from '../../types'
+import { TWidget } from '../types'
 
 Icon.add({ error_filled })
