--- packages/dm-core-plugins/src/form/widgets/TextareaWidget/index.ts
@@ -1,2 +1,0 @@
-export { default } from './TextareaWidget'
-export * from './TextareaWidget'
