--- packages/dm-core-plugins/src/form/widgets/Widgets.ts
@@ -1,9 +1,0 @@
-import CheckboxWidget from './CheckboxWidget/CheckboxWidget'
-import TextWidget from './TextWidget/TextWidget'
-import TextareaWidget from './TextareaWidget/TextareaWidget'
-
-export default {
-  CheckboxWidget,
-  TextWidget,
-  TextareaWidget,
-}
