+++ packages/dm-core-plugins/src/form/widgets/index.ts
@@ -1,2 +1,14 @@
-export { default } from './Widgets'
-export * from './Widgets'
+import { TWidget } from '../types'
+import CheckboxWidget from './CheckboxWidget'
+import TextWidget from './TextWidget'
+import TextareaWidget from './TextareaWidget'
+
+const widgets: {
+  [key: string]: (props: TWidget) => JSX.Element
+} = {
+  CheckboxWidget,
+  TextWidget,
+  TextareaWidget,
+}
+
+export default widgets
