--- packages/dm-core-plugins/src/form/widgets/index.tsx
@@ -1,7 +1,0 @@
-import TextWidget from './TextWidget'
-import CheckboxWidget from './CheckboxWidget'
-
-export default {
-  TextWidget,
-  CheckboxWidget,
-}
